third_party/blink/web_tests/wpt_internal/bluetooth/characteristic/notifications/stop-after-start-succeeds.https.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/bluetooth/resources/bluetooth-test.js"></script>
<script src="/bluetooth/resources/bluetooth-fake-devices.js"></script>
<script>
'use strict';
bluetooth_test(() => {
  return setBluetoothFakeAdapter('HeartRateAdapter')
    .then(() => requestDeviceWithTrustedClick({
      filters: [{services: ['heart_rate']}]}))
    .then(device => device.gatt.connect())
    .then(gattServer => gattServer.getPrimaryService('heart_rate'))
    .then(service => service.getCharacteristic('heart_rate_measurement'))
    .then(characteristic => {
      return characteristic.startNotifications()
        .then(start_characteristic => {
            assert_equals(start_characteristic, characteristic,
                'Start characteristic should the same as characteristic.');
            return characteristic.stopNotifications()
              .then(stop_characteristic => {
                assert_equals(stop_characteristic, start_characteristic,
                    'Stop characteristic should the same as start characteristic.');
              });
         });
     });
  // TODO(ortuno): Assert that notifications are not active.
  // http://crbug.com/600762
}, 'Single stop after start notifications succeeds.');
</script>
